Fixes map labels, projection improve, better docs
This version updates the time labels on the maps, namely the
forecast hours. This is done in the plot_utils.baseLabel function
where now the .total_seconds() method is used on the timedelta
between the forecast and the initialization time.

Projection improvements have been made in the form of keyword
arg parsing. Now, the individual plotting function assume that
the input x/y values are in the map projection if no transform
keyword is supplied; this is how GeoAxes work normally. If a
transform keyword is supplied, the x/y values are transformed
before any  plots are created. This speeds up plotting as
re-projection is only done once in the function instead of
for every plot. This setup will also allow even more speed up
because data could be projected once before any plots are made,
reducing the number of re-projections to one for all products.

Adds the xy_transform function to plot_utils to aid in
re-projection. This function strips off the units of the
x- and y-values (if they have units), projects the values
to the GeoAxes projection, and returns the re-projected
x- and y-values

Documentation in the functions has been improved, namely in
the header comments for inputs/outputs/keywords.

To do:
  - Whatever is left from v0.0.1
  - Create plotting code for rest of NAM products
    * Make general enough to use with any model
  - Update NAM download code to get variables required
    for other products.
  - Move .json files to top level and figure out how to do
    relative file paths for loading