Works on getting all NAM40 products working
Much work still need, but adds creation of single page plots
for each of the panels in the 4-panel plot. Must adjust the
layout for the single page plots and play with the scaling
of the base map some more.

Must also change how wind barbs are plotted as the density
of plots greatly increased on the single page plots; should
remain consistent throughout all plots.

For scaling issue/ perhaps could adjust scaling based on
maximum longitudinal extent of data; i.e., scale map so that
the model points furtherest south with be at the western/eastern
edges of the map domain.