Fixes colorbar size and map extent
This version fixes colorbar height issues that arose from scaling
the height based on height of axes; when full page axes,
colorbar too tall. Now, the height of a capital 'X' is used to
set the height of the colorbar based on the colorbar fontsize.
Width is still set based on the width of the axes.

The map extent is not determined mainly by the x- & y-values
of the data. In the getMapExtentScale function in plot_utils.py,
a scaling for the x- and y-axes is determine based on the axes
size and the x- and y-values. The larger of the two scales is
then used as the map scaling. Using this scaling, the map extent
is set to fill the entire map axes. The user can override the
auto scaling, however, this is discouraged.

Documention on functions has been updated.

To Do:
  - Create functions for the rest of the NAM forecast plots
  - Create a general function to create all the NAM forecast
    plots
      * This function should be adaptable to GFS, RAP, etc.
      * Possibly remove the forecast_4_panel function and
        place in the general function.
  - May also be a good idea to change various plotting functions
    to accept entire data dictionary as input and then use
    only those tags required.
  - Better centralization of contour interval definitions