Plotting functions now accept full dict
Full data dictionary is not input into the plotting function.
The functions then check to make sure the necessary varaibles
are present for plotting the various fields. If a var is NOT
present, then that layer is NOT plotted; error is logged.

Removed the forecast_4_panel function code as was not needed
now that other functions are cleaner (take dict and do checks).

In the NAM40_Products.py code, changed to creating one Figure()
instances that has axes added/cleared from while plotting; done
because was getting too many figures open warning. So now, figure
with page size and DPI is set up once, and then each time go to
plot, the figure is cleared and axes are added. However this
created warning with 'same input arguments returns already
created axes' warning, which caused plots to appear on-top-of
eachother. To solve this, the 'uuid.uuid4()' function is used
to generate a unique label for each axes (also added in the
add_colorbar plot_util) so that new one is always returned.

To Do
  - Still need to write functions for other products.
  - Convert core of NAM40_Products code to a general function