New time filtering working for awips models
Changs proposed in from v0.0.9 for getting model forecast times
is implimented. Then awips_fcst_times() function now accepts an
interval keyword for number of seconds between forecast data
to download. a max_forecast keyword is also added to limit the
length of the forecast data to download; time also in seconds.

All available times are iterated over, checking to see if the
forecast time falls on the requested interval. If it does, then
check to see if instantaneous period; i.e.,
time.getValidPeriod().duration() is zero (0), OR if the same
duration matches the requested interval. If either is True, then
index is determined based on forecast time and interval requested
and Datatime obeject is appened to appropriate nested list. Only
the nested list is returned by the object.

As function now returns nested list, then awips_model_base()
function time argument has been 'altered' in that it now expects
a list as input. The idea is that one will be iterating over the
nested list of times from awips_fcst_time, which will give lists
of times that can simply be input into awips_model_base().

v0.0.10