Significant changes to API, adds threaded download
The main model plotter code has been updated to be a class with
methods to make individual plots (4-panel, 1000-hPa, etc.), a
'main' plotter method to run all the other plotting methods,
and methods to download/generate plots for NAM and GFS.

Updated color_maps.py and contour_levels.py to better match keys/
height levels from the AWIPS download data.

Have also updated the AWIPSDownloader to use a thread for downloading.
Now, when using the getData() method, a generator is returned. In
the background, a thread is started that iterates over and downloads
data for given times, placing the downloaded data into a queue. The
queue will allow for 2 downloads to sit at a time; i.e., 2 items
in the queue, with a third set of downloaded data waiting to be
enqueued. The generator waits on the queue and returns data as it
becomes available. THe idea behind this that while data are being
plotted, which could take some time, data are still being downloaded
in the background.

v0.1.0