Enbales EDEX change at command line, bug fix
Have added the ability to specify the EDEX server at the command line.
Also added the ability to specify an output directory for the images.

There was a small bug in the date filtering where empty lists
could be passed to a method, and subsequent indexing of those
empty lists would cause an exception. I have wrapped the call
to said method in a try except so that it fails silently; well
there is debug logging so that I remember to handle this better
in the future....eventually

v0.1.1